feat: add all custom script to customScript file
